Update tests to use BizStream.Kentico.Xperience.AspNetCore.Mvc.Testing
