+++ nuget.config
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <packageSources>
+        <add key="local" value="/NugetServer" />
+    </packageSources>
+</configuration>
