+++ src/AspNetCore/src/XperienceStatusCodePagesExtensions.cs
@@ -39,7 +39,7 @@ namespace BizStream.Kentico.Xperience.AspNetCore.StatusCodePages
             => UseXperienceStatusCodePagesWithReExecute( app );
 
         /// <summary> Adds a StatusCodePages middleware to the pipeline. Specifies that responses should be handled by redirecting to a document in the Xperience Content Tree. </summary>
-        public static IApplicationBuilder UseXperienceStatusCodePagesWithRedirect( this IApplicationBuilder app )
+        public static IApplicationBuilder UseXperienceStatusCodePagesWithRedirects( this IApplicationBuilder app )
         {
             if( app == null )
             {
