+++ src/AspNetCore/test/BizStream.Kentico.Xperience.AspNetCore.StatusCodePages.Tests.csproj
@@ -11,8 +11,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <!--<PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="5.0.6" Condition=" '$(TargetFramework)' == 'net5.0' " />-->
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.15" Condition=" '$(TargetFramework)' == 'netcoreapp3.1' " />
+        <PackageReference Include="BizStream.Kentico.Xperience.AspNetCore.Mvc.Testing" Version="1.0.0" />
     </ItemGroup>
 
     <ItemGroup>
