+++ src/AspNetCore/test/StatusCodePageRetrieverTests.cs
@@ -1,17 +1,27 @@
 using System.Threading.Tasks;
+using BizStream.Kentico.Xperience.AspNetCore.Mvc.Testing;
 using BizStream.Kentico.Xperience.AspNetCore.StatusCodePages.Tests.Abstractions;
 using CMS.Core;
-using CMS.SiteProvider;
 using Kentico.Content.Web.Mvc;
 using NUnit.Framework;
 
 namespace BizStream.Kentico.Xperience.AspNetCore.StatusCodePages.Tests
 {
 
-    [TestFixture( Category = "Integration" )]
+    [TestFixture( Category = "IsolatedMvc" )]
     [TestOf( typeof( StatusCodePageRetriever ) )]
-    public class StatusCodePageRetrieverTests : BaseWebIntegrationTests
+    public class StatusCodePageRetrieverTests : StatusCodePagesTests<StatusCodePagesTestsStartup>
     {
+
+        protected override XperienceWebApplicationFactory<StatusCodePagesTestsStartup> CreateWebApplicationFactory( )
+        {
+            var factory = base.CreateWebApplicationFactory();
+
+            // NOTE: ensures current site context is preserved
+            factory.Server.PreserveExecutionContext = true;
+
+            return factory;
+        }
 
         [Test]
         [TestCase( 401 )]
@@ -29,10 +39,8 @@ namespace BizStream.Kentico.Xperience.AspNetCore.StatusCodePages.Tests
         }
 
         [SetUp]
-        public void StatusCodePageRetrieverSetUp( )
-        {
-            SiteContext.CurrentSiteName = "NewSite";
-        }
+        public async Task StatusCodePageRetrieverTestsSetUp( )
+            => await Client.GetAsync( "/" );
 
     }
